Store instagram image info in json
Former-commit-id: f442418eb4dd949132df5ea1713032a6ae87eed7

+++ scripts/instagrabber.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 
 import errno
+import json
 import logging
 import os
 import requests
@@ -17,6 +18,7 @@ output_dir = "assets/instagram_photos"
 
 
 output_dir = os.path.abspath(output_dir)
+output_json = output_dir + ".json"
 
 try:
     os.makedirs(output_dir)
@@ -42,13 +44,24 @@ def get_images():
             pass  # fuck videos man
 
         # Name file after media id
-        filename = media.id + ".jpg"
-        output_file = os.path.join(output_dir, filename)
-        if not os.path.exists(output_file):
-            logging.info("Instagrabber: Downloading {}...".format(filename))
+        output_file = os.path.join(output_dir, media.id + ".jpg")
+        output_json = os.path.join(output_dir, media.id + ".json")
+        if not (os.path.exists(output_file) and os.path.exists(output_json)):
+            logging.info("Instagrabber: Downloading {}...".format(media.id))
             r = requests.get(media.images['standard_resolution'].url)
             with open(output_file, 'wb') as f:
                 f.write(r.content)
+
+            # JSON data
+            file_info = json.dumps(
+                {
+                    # "caption": media.caption,
+                    "filename": output_file,
+                    "tags": [tag.name for tag in media.tags]
+                }
+            )
+            with open(output_json, 'wb') as j:
+                j.write(file_info)
 
 
 if __name__ == "__main__":
